Fix a bizarre hanging bug on os x mavericks (10.9) by not importing tkinter.
The following small program reproduces the hang on OS X 10.9:

---
import multiprocessing, sqlite3, _tkinter
def hang():
   sqlite3.connect('/tmp/foo')
if __name__ == '__main__':
   multiprocessing.Pool(2).apply_async(hang, []).get(999)
---

If you remover the '_tkinter' import then this code does *not* hang.

Who knows where this bug is?  But we don't use tkinter, so we don't
really care.

Auditors: chris, alpert

Test Plan:
Ran a version of the above program that imported pympler.muppy
(instead of _tkinter) before and after this change; it hangs before
but not after.